move DistributionFunctionFromFile into external module
